more use int64_t

+++ sperl_op.c
@@ -528,7 +528,7 @@ void SPerl_OP_resolve_types(SPerl* sperl) {
   
   SPerl_ARRAY* op_types = parser->op_types;
   
-  for (size_t i = 0, len = op_types->length; i < len; i++) {
+  for (int64_t i = 0, len = op_types->length; i < len; i++) {
     SPerl_OP* op_type = SPerl_ARRAY_fetch(sperl, op_types, i);
     _Bool success = SPerl_TYPE_resolve_type(sperl, op_type, 0);
     
@@ -551,14 +551,14 @@ void SPerl_OP_check(SPerl* sperl) {
   
   // Resolve package
   SPerl_ARRAY* op_packages = sperl->parser->op_packages;
-  for (size_t package_pos = 0; package_pos < op_packages->length; package_pos++) {
+  for (int64_t package_pos = 0; package_pos < op_packages->length; package_pos++) {
     SPerl_OP* op_package = SPerl_ARRAY_fetch(sperl, op_packages, package_pos);
     SPerl_PACKAGE* package = op_package->uv.package;
     SPerl_ARRAY* op_fields = package->op_fields;
     
     // Alignment is max size of field
     int32_t alignment = 0;
-    for (size_t field_pos = 0; field_pos < op_fields->length; field_pos++) {
+    for (int64_t field_pos = 0; field_pos < op_fields->length; field_pos++) {
       SPerl_OP* op_field = SPerl_ARRAY_fetch(sperl, op_fields, field_pos);
       SPerl_FIELD* field = op_field->uv.field;
       
@@ -571,7 +571,7 @@ void SPerl_OP_check(SPerl* sperl) {
     
     // Calculate package byte size
     int32_t package_byte_size = 0;
-    for (size_t field_pos = 0; field_pos < op_fields->length; field_pos++) {
+    for (int64_t field_pos = 0; field_pos < op_fields->length; field_pos++) {
       SPerl_OP* op_field = SPerl_ARRAY_fetch(sperl, op_fields, field_pos);
       SPerl_FIELD* field = op_field->uv.field;
